Added motormoving indicator.

+++ backend/robotarm/src/main/java/com/sundtrack/robotarm/sequence/dto/StepDto.java
@@ -9,6 +9,6 @@ public record StepDto(
         String name,
         PoseDto pose,
         Interpolation interpolation,
-        double speed
+        long speed
 ) {
 }
